requirements.txt
pytest==6.2.5
selenium==4.1.3
allure-pytest==2.9.45
requests==2.27.1
